LSTM baseline/requirements_20260427_000557.txt
neuralhydrology==1.9.1
pandas
numpy
xarray
cftime
netcdf4
scikit-learn
pyyaml
duckdb
pyarrow
torch
tqdm
